Change MPI metrics to 64bit

+++ criticalPath.h
@@ -109,7 +109,7 @@ enum ClockContext {
 };
 
 struct MPI_COUNTS {
-  int send{0}, recv{0}, isend{0}, irecv{0}, coll{0}, icoll{0}, test{0}, wait{0},
+  uint64_t send{0}, recv{0}, isend{0}, irecv{0}, coll{0}, icoll{0}, test{0}, wait{0},
       pers{0}, probe{0};
   void add(const MPI_COUNTS &o) {
     send += o.send;
